+++ lib/views/category.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:news_app/models/article.dart';
-import 'package:news_app/views/article_view.dart';
+import 'package:news_app/views/article.dart';
 import 'package:news_app/helper/news.dart';
 
 class CategoryNews extends StatefulWidget {
